correct the bugs

+++ byte&co/gestionoffre/src/main/java/com/glop/gestionoffre/OffreServiceImpl.java
@@ -1,12 +1,7 @@
 package com.glop.gestionoffre;
 
-
-import com.glop.gestionoffre.Offre;
-import com.glop.gestionoffre.OffreRepository;
-import com.glop.gestionoffre.OffreService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 
 @Service
@@ -16,32 +11,46 @@ public class OffreServiceImpl implements OffreService {
     private OffreRepository offreRepository;
 
     @Override
-    public List<Offre> getAllOffres() {
-        return offreRepository.findAll();
+    public List<OffreDTO> getAllOffres() {
+        List<Offre> offres = offreRepository.findAll();
+        return OffreMapper.toDTOList(offres);
     }
 
     @Override
-    public Offre getOffreById(Long id) {
-        return offreRepository.findById(id).orElseThrow(() -> new RuntimeException("Offre not found"));
+    public OffreDTO getOffreById(Long id) {
+        Offre offre = offreRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Offre not found"));
+        return OffreMapper.toDTO(offre);
     }
 
     @Override
-    public Offre createOffre(Offre offre) {
-        return offreRepository.save(offre);
+    public OffreDTO createOffre(OffreDTO offreDTO) {
+        Offre offre = OffreMapper.fromDTO(offreDTO);
+        Offre createdOffre = offreRepository.save(offre);
+        return OffreMapper.toDTO(createdOffre);
     }
 
     @Override
-    public Offre updateOffre(Long id, Offre offre) {
-        Offre existingOffre = getOffreById(id);
-        existingOffre.setNomoffre(offre.getNomoffre());
-        existingOffre.setPrixoffre(offre.getPrixoffre());
-        existingOffre.setDescriptionoffre(offre.getDescriptionoffre());
-        existingOffre.setConditionseligibilite(offre.getConditionseligibilite());
-        return offreRepository.save(existingOffre);
+    public OffreDTO updateOffre(Long id, OffreDTO offreDTO) {
+        Offre existingOffre = offreRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Offre not found"));
+        
+        existingOffre.setNomoffre(offreDTO.getNomoffre());
+        //existingOffre.setPrixoffre(offreDTO.getPrixoffre());
+        existingOffre.setDescriptionoffre(offreDTO.getDescriptionoffre());
+        existingOffre.setConditionseligibilite(offreDTO.getConditionseligibilite());
+        existingOffre.setImageoffre(offreDTO.getImageoffre());
+        existingOffre.setAvantagesoffre(offreDTO.getAvantagesoffre());
+
+        Offre updatedOffre = offreRepository.save(existingOffre);
+        return OffreMapper.toDTO(updatedOffre);
     }
 
     @Override
     public void deleteOffre(Long id) {
+        if (!offreRepository.existsById(id)) {
+            throw new RuntimeException("Offre not found");
+        }
         offreRepository.deleteById(id);
     }
 }
